Fix fileserver.roots._file_lists for Windows symlinks
